api/requirements.txt
fastapi
pydantic-ai[openai]
pydantic-settings
python-dotenv
uvicorn
